publisher/pyproject.toml
[project]
name = "almanach-publisher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "msgpack>=1.1.2",
    "nats-py>=2.12.0",
    "rich>=14.2.0",
]

[project.scripts]
almanach-publisher = "almanach_publisher:main"

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"
